fix: move pagination to separate component
