+++ src/components/Category/ProductList/ProductList.tsx
@@ -2,6 +2,7 @@ import type { ProductProjection } from '@commercetools/platform-sdk'
 import s from '../Category.module.scss'
 import type { ReactElement } from 'react'
 import { Link } from 'react-router-dom'
+import { Pagination } from '@/components/Pagination/Pagination.tsx'
 
 type ProductListProperties = {
   currentPage: number
@@ -19,53 +20,9 @@ export const ProductList = ({
   products,
 }: ProductListProperties): ReactElement => {
   const CENTS_IN_DOLLAR = 100
-  const MAX_PAGE_BUTTONS = 4
 
   const totalProducts = total || 0
   const totalPages = Math.ceil(totalProducts / pageSize)
-
-  const getPaginationButtons = (): (number | string)[] => {
-    const buttons: (number | string)[] = []
-
-    if (totalPages <= MAX_PAGE_BUTTONS) {
-      return Array.from({ length: totalPages }, (_, pageIndex) => pageIndex)
-    }
-
-    const sideButtons = 2
-    const aroundCurrent = 2
-    buttons.push(0)
-
-    let start = Math.max(sideButtons, currentPage - aroundCurrent)
-    let end = Math.min(totalPages - sideButtons, currentPage + aroundCurrent)
-
-    const buttonsNeeded = MAX_PAGE_BUTTONS - sideButtons
-    if (end - start + 1 < buttonsNeeded) {
-      if (currentPage < totalPages / sideButtons) {
-        end = Math.min(totalPages - 1, start + buttonsNeeded - 1)
-      } else {
-        start = Math.max(sideButtons, end - buttonsNeeded + 1)
-      }
-    }
-
-    if (start > sideButtons) {
-      buttons.push('...')
-    }
-
-    for (let index = start - 1; index <= end; index++) {
-      buttons.push(index)
-    }
-
-    if (end < totalPages - sideButtons) {
-      buttons.push('...')
-    }
-
-    buttons.push(totalPages - 1)
-    return buttons
-  }
-
-  const goToPage = (page: number): void => {
-    onPageChange(Math.max(0, Math.min(page, totalPages - 1)))
-  }
 
   return (
     <section className={s.productssection}>
@@ -99,33 +56,7 @@ export const ProductList = ({
         })}
       </ul>
       {totalPages > 1 && (
-        <div className={s.pagination}>
-          <button onClick={() => goToPage(currentPage - 1)} disabled={currentPage === 0} className={s.paginationbutton}>
-            Previous
-          </button>
-          {getPaginationButtons().map((pageIndex, index) =>
-            typeof pageIndex === 'string' ? (
-              <span key={`ellipsis-${index}`} className={s.ellipsis}>
-                {pageIndex}
-              </span>
-            ) : (
-              <button
-                key={pageIndex}
-                onClick={() => goToPage(pageIndex)}
-                className={`${s.paginationbutton} ${currentPage === pageIndex ? s.active : ''}`}
-              >
-                {pageIndex + 1}
-              </button>
-            )
-          )}
-          <button
-            onClick={() => goToPage(currentPage + 1)}
-            disabled={currentPage === totalPages - 1}
-            className={s.paginationbutton}
-          >
-            Next
-          </button>
-        </div>
+        <Pagination totalPages={totalPages} currentPage={currentPage} onPageChange={onPageChange}></Pagination>
       )}
     </section>
   )
